backend/pyproject.toml
[project]
name = "after_school_backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.6",
    "strawberry-graphql[fastapi]>=0.246.1",
    "sqlalchemy>=2.0.35",
    "psycopg[binary]>=3.2.1",
    "alembic>=1.13.2",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "redis>=5.0.8",
    "types-redis>=4.6.0",
    "asyncpg>=0.31.0",
    "auth0-fastapi-api>=1.0.0b5",
    "python-jose>=3.5.0",
    "cryptography>=46.0.3",
    "httpx>=0.28.1",
    "aiodataloader>=0.4.3",
    "colorlog>=6.10.1",
]
